requirements.txt
requests
dotenv
PyJWT